add iris example

+++ src/api.jl
@@ -8,7 +8,7 @@
 # ## COMPULSORY
 
 # train `model` for another `Δn` iterations:
-train!(model, Δn::Integer) =
+train!(model, Δn) =
     error("`IterationControl.train!(model, ::Int)` not overloaded "*
           "for `typeof(model) = $(typeof(model))`. ")
